quantization unit in bandpass filter

+++ packages/mountainsort2/src/p_split_firings.cpp
@@ -6,9 +6,9 @@
 namespace P_split_firings {
 
 struct Session {
-    int start_timepoint = 0;
-    int end_timepoint = 0;
-    QVector<int> event_indices;
+    bigint start_timepoint = 0;
+    bigint end_timepoint = 0;
+    QVector<bigint> event_indices;
 };
 }
 
@@ -17,11 +17,11 @@ bool p_split_firings(QStringList timeseries_paths, QString firings_path, QString
     if (timeseries_paths.count() == 0)
         return true;
 
-    QVector<int> start_timepoints;
+    QVector<bigint> start_timepoints;
 
     QList<P_split_firings::Session> sessions;
-    int offset = 0;
-    for (int i = 0; i < timeseries_paths.count(); i++) {
+    bigint offset = 0;
+    for (bigint i = 0; i < timeseries_paths.count(); i++) {
         DiskReadMda32 X(timeseries_paths[i]);
         P_split_firings::Session S;
         S.start_timepoint = offset;
@@ -31,8 +31,8 @@ bool p_split_firings(QStringList timeseries_paths, QString firings_path, QString
     }
 
     Mda firings(firings_path);
-    int current_timeseries_ind = 0;
-    for (int i = 0; i < firings.N2(); i++) {
+    bigint current_timeseries_ind = 0;
+    for (bigint i = 0; i < firings.N2(); i++) {
         double time0 = firings.value(1, i);
         while ((time0 < sessions[current_timeseries_ind].start_timepoint) && (current_timeseries_ind - 1 >= 0)) {
             current_timeseries_ind--;
@@ -43,10 +43,10 @@ bool p_split_firings(QStringList timeseries_paths, QString firings_path, QString
         sessions[current_timeseries_ind].event_indices << i;
     }
 
-    for (int i = 0; i < sessions.count(); i++) {
+    for (bigint i = 0; i < sessions.count(); i++) {
         Mda firings0(firings.N1(), sessions[i].event_indices.count());
-        for (int j = 0; j < sessions[i].event_indices.count(); j++) {
-            for (int r = 0; r < firings.N1(); r++) {
+        for (bigint j = 0; j < sessions[i].event_indices.count(); j++) {
+            for (bigint r = 0; r < firings.N1(); r++) {
                 firings0.setValue(firings.value(r, sessions[i].event_indices[j]), r, j);
             }
             double time0 = firings0.value(1, j);
